refactor: move auth and avatar services to gn4 repository libs/api
