+++ libs/api/repository/src/lib/gn4/auth/avatar.service.interface.ts
