+++ libs/api/repository/src/lib/gn4/auth/gravatar.service.spec.ts
@@ -1,8 +1,7 @@
 import { TestBed } from '@angular/core/testing'
-
 import { GravatarService } from './gravatar.service'
 import { HttpClientTestingModule } from '@angular/common/http/testing'
-import { Gn4SettingsService } from '@geonetwork-ui/api/repository/gn4'
+import { Gn4SettingsService } from '../settings/gn4-settings.service'
 import { BehaviorSubject } from 'rxjs'
 
 class Gn4SettingsServiceMock {
