+++ libs/api/repository/src/lib/gn4/index.ts
@@ -2,3 +2,5 @@ export * from './gn4-repository'
 export * from './organizations'
 export * from './elasticsearch'
 export * from './settings/gn4-settings.service'
+export * from './auth'
+export * from './favorites/favorites.service'
