+++ libs/feature/auth/src/index.ts
@@ -1,4 +1,1 @@
 export * from './lib/feature-auth.module'
-export * from './lib/auth.service'
-export * from './lib/avatar/avatar.service.interface'
-export * from './lib/avatar/gravatar.service'
