+++ libs/feature/auth/src/lib/feature-auth.module.ts
@@ -1,7 +1,9 @@
 import { NgModule } from '@angular/core'
 import { CommonModule } from '@angular/common'
-import { AvatarServiceInterface } from './avatar/avatar.service.interface'
-import { GravatarService } from './avatar/gravatar.service'
+import {
+  AvatarServiceInterface,
+  GravatarService,
+} from '@geonetwork-ui/api/repository/gn4'
 
 @NgModule({
   imports: [CommonModule],
